Code restructure
Split code tree into directories and smaller files to make it easier to
maintain - Phase 1
