+++ examples/connect-cluster.py
@@ -1,9 +1,9 @@
-from scaleiopy import scaleio
+from scaleiopy.scaleio import ScaleIO
 from pprint import pprint
 import time
 import sys
 
-sio = scaleio.ScaleIO("https://" + sys.argv[1] + "/api",sys.argv[2], sys.argv[3], False, "ERROR") # HTTPS must be used as there seem to be an issue with 302 responses in Requests when using POST
+sio = ScaleIO("https://" + sys.argv[1] + "/api",sys.argv[2], sys.argv[3], False, "ERROR") # HTTPS must be used as there seem to be an issue with 302 responses in Requests when using POST
 print "{}{}".format("--- Current ScaleIO API version is: ", sio.get_api_version())
 print "--- ScaleIO System ---"
 pprint(sio.system)
