+++ examples/create-snapshot-of-volume.py
@@ -1,16 +1,18 @@
-from scaleiopy import *
+from scaleiopy.scaleio import ScaleIO
+from scaleiopy.api.scaleio.mapping.snapshotspecification import SnapshotSpecification
 from pprint import pprint
 import sys
 
 # How to run:
 # python create-snapshot-of-volume.py ip-to-gw user pass volume_name
 
-sio = scaleio.ScaleIO("https://" + sys.argv[1] + "/api",sys.argv[2],sys.argv[3],False,"ERROR") # HTTPS must be used as there seem to be an issue with 302 responses in Requests when using POST
+sio = ScaleIO("https://" + sys.argv[1] + "/api",sys.argv[2],sys.argv[3],False,"ERROR") # HTTPS must be used as there seem to be an issue with 302 responses in Requests when using POST
     
 #sio.create_volume_by_pd_name(sys.argv[4], 8192, sio.get_pd_by_name(sys.argv[5]), mapAll=True)
 #pprint(sio.volumes)
 
-snapSpec = scaleio.SnapshotSpecification()
+
+snapSpec = SnapshotSpecification()
 snapSpec.addVolume(sio.get_volume_by_name(sys.argv[4]))
 print "**********"
 print "* Volume *"
