+++ examples/create-volume.py
@@ -1,4 +1,4 @@
-from scaleiopy import scaleio
+from scaleiopy.scaleio import ScaleIO
 from pprint import pprint
 import sys
 
@@ -8,7 +8,7 @@ import sys
 # Whats this code doing:
 # Create a volume of x MB inside specified protectiondomain and storagepool
 
-sio = scaleio.ScaleIO("https://" + sys.argv[1] + "/api",sys.argv[2],sys.argv[3],False,"ERROR") # HTTPS must be used as there seem to be an issue with 302 responses in Requests when using POST
+sio = ScaleIO("https://" + sys.argv[1] + "/api",sys.argv[2],sys.argv[3],False,"ERROR") # HTTPS must be used as there seem to be an issue with 302 responses in Requests when using POST
     
 sio.create_volume(sys.argv[4], sys.argv[5], sio.get_pd_by_name(sys.argv[6]), sio.get_storage_pool_by_name(sys.argv[7]))
 pprint(sio.volumes)
