+++ examples/unmap_delete-volume.py
@@ -1,4 +1,4 @@
-from scaleiopy import scaleio
+from scaleiopy.scaleio import ScaleIO
 from pprint import pprint
 import sys
 
@@ -8,6 +8,6 @@ import sys
 # Whats this code doing:
 # Deletes specified volume and unmap from all mapped SDCs
 
-sio = scaleio.ScaleIO("https://" + sys.argv[1] + "/api",sys.argv[2],sys.argv[3],False,"ERROR") # HTTPS must be used as there seem to be an issue with 302 responses in Requests when using POST
+sio = ScaleIO("https://" + sys.argv[1] + "/api",sys.argv[2],sys.argv[3],False,"ERROR") # HTTPS must be used as there seem to be an issue with 302 responses in Requests when using POST
 
 sio.delete_volume(sio.get_volume_by_name(sys.argv[4]), 'ONLY_ME', autoUnmap=True)
