+++ scaleiopy/__init__.py
@@ -1,1 +1,6 @@
-from scaleio import *
+# Imports
+
+#from api.im import im_generic_object
+#import im_generic_object as Im_Generic_Object
+#from api.im import system #ScaleIO_System_Object
+
