+++ scaleiopy/api/im/mapping/mdm.py
@@ -1,0 +1,33 @@
+# Imports
+from im_generic_object import Im_Generic_Object
+from node import Node_Object
+
+class Mdm_Object(Im_Generic_Object):
+    """
+    Python object representation of a MDM (primary or secondary look eactly the same configuration wise).
+    """
+    
+    def __init__(self,
+        node=None,
+        nodeInfo=None,
+        managementIPs=None,
+        mdmIPs = None
+    ):
+        self.managementIPs=[]
+        if managementIPs:
+            for mgmtIP in managementIPs:
+                self.managementIPs.append(mgmtIP)
+        self.node=Node_Object.from_dict(node)
+        self.nodeInfo=nodeInfo
+        self.mdmIPs = []
+        if mdmIPs:
+            for mdmIP in mdmIPs:
+                self.mdmIPs.append(mdmIP)
+                
+    @staticmethod
+    def from_dict(dict):
+        """
+        A convinience method that directly creates a new instance from a passed dictionary (that probably came from a
+        JSON response from the server.
+        """
+        return Mdm_Object(**dict)
