+++ scaleiopy/api/scaleio/__init__.py
@@ -1,0 +1,23 @@
+# Imports
+
+
+class SIO_Generic_Object(object):
+    @classmethod
+    def get_class_name(cls):
+        """
+        A helper method that returns the name of the class.  Used by __str__ below
+        """
+        return cls.__name__
+
+    def __str__(self):
+        """
+        A convenience method to pretty print the contents of the class instance
+        """
+        # to show include all variables in sorted order
+        return "<{}> @ {}:\n".format(self.get_class_name(), id(self)) + "\n".join(
+            ["  %s: %s" % (key.rjust(22), self.__dict__[key]) for key in sorted(set(self.__dict__))])
+
+    def __repr__(self):
+        return self.__str__()
+    
+
