+++ scaleiopy/util/installerfsm.py
@@ -1,11 +1,11 @@
+from pprint import pprint
 import json
 import time
-
-from scaleioutil import ScaleIOLogger
 
 
 """
 TODO:
+* Add logging (replace all print statements) - Use self.FSM.imapi.logger.debug() to access IM objects existing logger facility
 
 """
 
@@ -23,7 +23,8 @@ class Transition(object):
 		self.toState = toState
 		
 	def Execute(self):
-		self.logger.debug("Transitioning")
+		print ("Transitioning...")
+		print ""
 
 
 ##===============================================
@@ -37,6 +38,7 @@ class State(object):
 		self.startTime = 0
 		im_instance = im_inst
 		# Possible states: IDLE, PENDING, FAILED, COMPLETED
+		#print self.FSM.getCurrentStateStatus()
 		
 	def Enter(self):
 		self.FSM.setCurrentStateStatus('PENDING')
@@ -61,68 +63,68 @@ class Query(State):
 		self.calling_object = im_inst
 		
 	def Enter(self):
-		self.logger.debug("Enter Query State")
+		print ("Entering QUERY phase")
 		super(Query, self).Enter()
 
 	def Execute (self):
-		self.logger.debug("Start ScaleIO IM Query phase")
+		print ("Start ScaleIO IM Query phase")
 		self.FSM.imapi.set_state('query')
-		self.logger.debug("State: {}".format(self.FSM.imapi.get_state()))
-
+		print self.FSM.imapi.get_state()
 		self.pendingItemCount = self.FSM.install_process_status("query")
 		while self.pendingItemCount > 0:
 			self.pendingItemCount = self.FSM.install_process_status("query")
 			time.sleep(0.5)
 		if self.pendingItemCount < 0:
-			self.logger.error("Query State Failed!")
+			print "*** FAILED ***"
 		else:
-			self.logger.debug("COMPLETED")
+			print "*** COMPLETED ***"
 			self.Next()
 
 		if self.FSM.autoTransition:
 			self.FSM.Execute()
 
 	def Next(self):
-		self.logger.debug("Advance to UPLOAD phase")
+		#print ("Advance to next step")
 		self.FSM.ToTransition("toUPLOAD") # Move to next step
 		
 	def Exit(self):
 		pass
+		#print ("Exiting QUERY phase.")
 
 class Upload(State):
-	# Worker class - Takes care of managing IM UPLOAD phase and do status control of each phase
+	# Worker class - Takes care of managing IM QUERY phase and do status control of each phase
 	def __init__(self, FSM, im_inst):
 		super(Upload, self).__init__(FSM, im_inst)
 		
 	def Enter(self):
-		self.logger.debug("Enter Upload State")
-
+		print ("Entering UPLOAD phase")
 		super(Upload, self).Enter()
 
 	def Execute (self):
-		self.logger.debug("Upload of binaries to nodes")
+		print ("Upload of binaries to nodes")
 		# Call set_state
 		self.FSM.imapi.set_state('upload') # Set IM to UPLOAD state
-		self.logger.debug("State: {}".format(self.FSM.imapi.get_state()))
+		print self.FSM.imapi.get_state()
 		self.pendingItemCount = self.FSM.install_process_status("upload")
 		while self.pendingItemCount > 0:
 			self.pendingItemCount = self.FSM.install_process_status("upload")
 			time.sleep(2)
 		if self.pendingItemCount < 0:
-			self.logger.error("Upload State Failed!")
+			print "*** FAILED ***"
 		else:
-			self.logger.debug("COMPLETED")
+			print "*** COMPLETED ***"
 			self.Next()
 		
 		if self.FSM.autoTransition:
 			self.FSM.Execute()
 	
 	def Next(self):
-		self.logger.debug("Advance to INSTALL phase")
+		#print ("Advance to next step")
 		self.FSM.ToTransition("toINSTALL") # Move to next step
 		
 	def Exit(self):
 		pass
+		#print ("Exiting UPLOAD phase.")
 
 
 class Install(State):
@@ -131,21 +133,21 @@ class Install(State):
 		super(Install, self).__init__(FSM, im_inst)
 		
 	def Enter(self):
-		self.logger.debug("Enter Install State")
+		print ("Entering install phase")
 		super(Install, self).Enter()
 
 	def Execute (self):
-		self.logger.debug("Installing ScaleIO binaries on nodes")
-		self.FSM.imapi.set_state('install') # Set IM to INSTALL state
-		self.logger.debug("State: {}".format(self.FSM.imapi.get_state()))
+		print ("Installing ScaleIO binaries")
+		self.FSM.imapi.set_state('install') # Set IM to QUERY state
+		print self.FSM.imapi.get_state()
 		self.pendingItemCount = self.FSM.install_process_status("install")
 		while self.pendingItemCount > 0:
 			self.pendingItemCount = self.FSM.install_process_status("install")
 			time.sleep(2)
 		if self.pendingItemCount < 0:
-			self.logger.error("FAILED")
+			print "*** FAILED ***"
 		else:
-			self.logger.debug("COMPLETED")
+			print "*** COMPLETED ***"
 			time.sleep(6)
 			self.Next()
 		
@@ -153,11 +155,11 @@ class Install(State):
 			self.FSM.Execute()
 			
 	def Next(self):
-		self.logger.debug("Advance to Configure Phase")
+		print ("Advance to next step")
 		self.FSM.ToTransition("toCONFIGURE") # Move to next step
 		
 	def Exit(self):
-		self.logger.debug("Exit INSTALL phase")
+		print ("Exiting INSTALL phase.")
 
 
 class Configure(State):
@@ -166,33 +168,34 @@ class Configure(State):
 		super(Configure, self).__init__(FSM, im_inst)
 		
 	def Enter(self):
-		self.logger.debug("Entering Configure Phase")
+		print ("Entering configure phase")
 		super(Configure, self).Enter()
 
 	def Execute (self):
-		self.logger.debug("Configure ScaleIO Cluster")
+		print ("Configure ScaleIO cluster")
 		# Call set_state
 		self.FSM.imapi.set_state('configure') # Set IM to QUERY state
-		self.logger.debug("State: {}".format(self.FSM.imapi.get_state()))
+		print self.FSM.imapi.get_state()
 		self.pendingItemCount = self.FSM.install_process_status("configure")
 		while self.pendingItemCount > 0:
 			self.pendingItemCount = self.FSM.install_process_status("configure")
 			time.sleep(2)
 		if self.pendingItemCount < 0:
-			self.logger.error("FAILED")
+			print "*** FAILED ***"
 		else:
-			self.logger.debug("COMPLETED")
+			print "*** COMPLETED ***"
 			self.Next()
 
 		if self.FSM.autoTransition:
 			self.FSM.Execute()
 		
 	def Next(self):
-		self.logger.debug("Advance to ARCHIVE State")
+		#print ("Advance to next step")
 		self.FSM.ToTransition("toARCHIVE") # Move to next step
 		
 	def Exit(self):
 		pass
+		#print ("Exiting CONFIGURE phase.")
 
 class Archive(State):
 	# Worker class - Takes care of managing IM Archive phase and do status control of each phase
@@ -200,13 +203,14 @@ class Archive(State):
 		super(Archive, self).__init__(FSM, im_inst)
 		
 	def Enter(self):
-		self.logger.debug("Entering Archive phase")
+		print ("Entering archive phase")
 		super(Archive, self).Enter()
 
 	def Execute (self):
-		self.logger.debug("Completing ScaleIO cluster installation")
+		print ("Completing ScaleIO cluster install")
+		#print ("Execute self.set_archive_all()")
 		self.FSM.imapi.set_archive_all()
-		self.logger.debug("State: {}".format(self.FSM.imapi.get_state()))
+		print self.FSM.imapi.get_state()
 		time.sleep(3)
 		
 		self.Next()
@@ -214,11 +218,11 @@ class Archive(State):
 			self.FSM.Execute()
 		
 	def Next(self):
-		self.logger.debug("Advance to COMPLETE step")
+		print ("Advance to next step")
 		self.FSM.ToTransition("toCOMPLETE") # Move to next step
 		
 	def Exit(self):
-		self.logger.debug("Exiting ARCHIVE phase")
+		print ("Exiting ARCHIVE phase.")
 
 class Complete(State):
 	# Worker class - Takes care of managing IM "Completing" phase and do status control of each phase
@@ -226,21 +230,20 @@ class Complete(State):
 		super(Complete, self).__init__(FSM, im_inst)
 		
 	def Enter(self):
-		self.logger.debug("Entering COMPLETE phase")
+		print ("Entering COMPLETE phase")
 		super(Complete, self).Enter()
 
 	def Execute (self):
-		self.logger.debug("Installation complete!")
+		print ("Installation complete!")
 		self.FSM.setCurrentStateStatus('COMPLETED')
-		self.logger.debug("Current State Status: {}".format(self.FSM.getCurrentStateStatus()))
-		self.logger.debug("State: {}".format(self.FSM.imapi.get_state()))
-
-
+		print self.FSM.getCurrentStateStatus()
+		print self.FSM.imapi.get_state()
+		
 	def Next(self):
 		self.FSM.setCurrentStateStatus('COMPLETED')
 
 	def Exit(self):
-		self.logger.debug("Exiting COMPLETE phase")
+		print ("Exiting COMPLETE phase.")
 
 ##===============================================
 ## FINITE STATE MACHINE
@@ -279,11 +282,12 @@ class FSM(object):
 		return curState
 	
 	def setCurrentStateStatus(self, status):
-		self.logger.debug("*** Setting currentStateStatus to {} ***".format(status))
+		print "*** Setting currentStateStatus to " + status + " ***" 
 		self.curStateStatus = status
 		if status == 'COMPLETED':
 			self.trans = None
 		#	self.curState = None
+		#print "setCurrentStateStatus() = " + self.curStateStatus
 	
 	def getCurrentStateStatus(self):
 		return self.curStateStatus
@@ -299,10 +303,10 @@ class FSM(object):
 				self.Execute()
 		else:
 			self.trans = None
-			self.logger.debug("Status is COMPLETE - Next() wont do anything")
+			print " Status is COMPLETE - Next() will not do anything"
 			
 	def Execute(self):
-		self.logger.debug("FSM Execute() - curStateStatus = ".format(self.curStateStatus))
+		print "FSM Execute() - curStateStatus = " + self.curStateStatus
 		if self.curStateStatus == 'IDLE' or self.curStateStatus == 'PENDING':
 			self.setCurrentStateStatus('PENDING')
 			if (self.trans):
@@ -313,7 +317,7 @@ class FSM(object):
 				self.trans = None
 			self.curState.Execute()
 		else:
-			self.logger.debug("Statemachine COMPLETE")
+			print "Statemachine COMPLETE"
 			self.trans = None
 
 	
@@ -329,7 +333,7 @@ class FSM(object):
 						return -1
 					if innerArray['commandState'] == 'pending':
 						self.pendingItemCount += 1
-		self.logger.debug("Pending Items: ".format(str(self.pendingItemCount)))
+		print "Pending Items = " + str(self.pendingItemCount)
 		if self.pendingItemCount == 0:
 			return 0
 		else:
@@ -349,12 +353,10 @@ class InstallerFSM:
 	# Base character which will be holding the Finite State Machine,
 	# which in turn will hold the states and transitions.
 	def __init__(self, imapi, automatic=False):
-		self.loggerInstance = ScaleIOLogger.get() # Create logger Singleton
-		self.logger = self.loggerInstance.getLoglevel(debugLevel) # Get logger instance with defined logging level
 		self.FSM = FSM(imapi)
 		if automatic:
 			self.FSM.enableAutoTransition()
-
+			
 		## STATES
 		self.FSM.AddState("QUERY", Query(self.FSM, imapi))
 		self.FSM.AddState("UPLOAD", Upload(self.FSM, imapi))
